requirements.txt
lancedb>=0.4.0
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1
langchain-google-genai>=2.1.0
google-ai-generativelanguage>=0.6.0
pandas>=1.5.0
docling>=0.1.0
torch>=2.0.0
transformers>=4.30.0
psycopg2-binary>=2.9.0
requests>=2.28.0
python-dotenv>=1.0.0 
flask>=3.0.0